+++ client/ayon_nuke/plugins/publish/intgegrate_prores.py
@@ -1,0 +1,13 @@
+# import pyblish.api
+
+
+# class IntegrateProres(pyblish.api.InstancePlugin):
+
+#     label = "Integrate untracked prores"
+#     order = pyblish.api.IntegratorOrder + 0.52
+#     hosts = ['nuke']
+#     families = ["render",'plate','prerender']
+#     # default_template_name = "publish"
+
+#     def process(self, instance):
+#         self.log.info("Integrating untracked prores")
